refactor lesson component to file and dynamically import
